+++ core/server/api_container/server/service_network/mock_service_network.go
@@ -32,7 +32,7 @@ func (m MockServiceNetwork) RegisterService(serviceId service_network_types.Serv
 	panic("This is unimplemented for the mock network")
 }
 
-func (m MockServiceNetwork) StartService(ctx context.Context, serviceId service_network_types.ServiceID, imageName string, privatePorts map[string]*enclave_container_launcher.EnclaveContainerPort, entrypointArgs []string, cmdArgs []string, dockerEnvVars map[string]string, enclaveDataDirMountDirpath string, filesArtifactMountDirpaths map[string]string) (resultPublicIpAddr net.IP, resultPublicPorts map[string]*enclave_container_launcher.EnclaveContainerPort, resultErr error) {
+func (m MockServiceNetwork) StartService(ctx context.Context, serviceId service_network_types.ServiceID, imageName string, privatePorts map[string]*enclave_container_launcher.EnclaveContainerPort, entrypointArgs []string, cmdArgs []string, dockerEnvVars map[string]string, enclaveDataDirMountDirpath string, filesArtifactMountDirpaths map[string]string) (resultMaybePublicIpAddr net.IP, resultPublicPorts map[string]*enclave_container_launcher.EnclaveContainerPort, resultErr error) {
 	panic("This is unimplemented for the mock network")
 }
 
@@ -52,7 +52,7 @@ func (m MockServiceNetwork) GetServiceRegistrationInfo(serviceId service_network
 	return ip, "", nil
 }
 
-func (m MockServiceNetwork) GetServiceRunInfo(serviceId service_network_types.ServiceID) (privatePorts map[string]*enclave_container_launcher.EnclaveContainerPort, publicIpAddr net.IP, publicPorts map[string]*enclave_container_launcher.EnclaveContainerPort, enclaveDataDirMntDirpath string, resultErr error) {
+func (m MockServiceNetwork) GetServiceRunInfo(serviceId service_network_types.ServiceID) (privatePorts map[string]*enclave_container_launcher.EnclaveContainerPort, maybePublicIpAddr net.IP, publicPorts map[string]*enclave_container_launcher.EnclaveContainerPort, enclaveDataDirMntDirpath string, resultErr error) {
 	dataDirMntDirpath, found := m.serviceEnclaveDataDirMntDirpaths[serviceId]
 	if !found {
 		return nil, nil, nil, "", stacktrace.NewError("No enclave data directory mount dirpath defined for service with ID '%v'", serviceId)